Code/backend/requirements.txt
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
transformers>=5.0.0
torch>=2.0.0
torchvision>=0.15.0
huggingface_hub>=0.20.0
Pillow>=10.0.0
datasets>=4.0.0
